move func parsing to aqua-run
