+++ cli/.js/src/main/scala/aqua/ArgOpts.scala
@@ -32,64 +32,12 @@ case class VarJson(variable: VarRaw, value: js.Dynamic)
 
 object ArgOpts {
 
-  def spanToId: Span.S ~> Id = new (Span.S ~> Id) {
-
-    override def apply[A](span: Span.S[A]): Id[A] = {
-      span._2
-    }
-  }
-
   // Parses a function name and arguments from a string
   def funcOpt: Opts[CliFunc] =
     Opts
       .option[String]("func", "Function to call with args", "f", "funcName(args)")
       .mapValidated { str =>
-        CallArrowToken.callArrow.parseAll(str.trim) match {
-          case Right(exprSpan) =>
-            val expr = exprSpan.mapK(spanToId)
-
-            val argsV = expr.args.collect {
-              case LiteralToken(value, ts) =>
-                validNel(LiteralRaw(value, ts))
-              case VarToken(name, _) =>
-                validNel(VarRaw(name.value, BottomType))
-              case CollectionToken(_, values) =>
-                val hasVariables = values.exists {
-                  case LiteralToken(_, _) => false
-                  case _ => true
-                }
-                if (!hasVariables) {
-                  val literals = values.collect { case LiteralToken(value, ts) =>
-                    LiteralRaw(value, ts)
-                  }
-                  val hasSameTypesOrEmpty =
-                    literals.isEmpty || literals.map(_.baseType).toSet.size == 1
-
-                  if (hasSameTypesOrEmpty) {
-                    validNel(
-                      NonEmptyList
-                        .fromList(literals)
-                        .map(l => CollectionRaw(l, ArrayType(l.head.baseType)))
-                        .getOrElse(ValueRaw.Nil)
-                    )
-                  } else
-                    invalidNel(
-                      "If the argument is an array, then it must contain elements of the same type."
-                    )
-
-                } else
-                  invalidNel(
-                    "Array arguments can only have numbers, strings, or booleans."
-                  )
-              case CallArrowToken(_, _, _) =>
-                invalidNel("Function calls as arguments are not supported.")
-            }.sequence
-            argsV.andThen(args =>
-              validNel(CliFunc(expr.funcName.value, args, expr.ability.map(_.name)))
-            )
-
-          case Left(err) => invalid(err.expected.map(_.context.mkString("\n")))
-        }
+        CliFunc.fromString(str)
       }
 
   // Gets data from a file or from a json string
